moving store out of util
