--- src/hat_util/mod.rs
@@ -1,3 +1,0 @@
-mod store;
-
-pub use store::{Store, StoreComposed, StoreUnion};
